6. WEB/Исходники тасок/Advanced/JawaToken/src/requirements.txt
flask == 3.1.*
pyjwt == 2.10.*
